api/requirements.txt
fastapi
uvicorn[standard]
pydantic
celery[redis]
geojson
psycopg2-binary
